requirements.txt
tensorflow==2.20.0
streamlit==1.39.0
pillow==10.4.0
numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.2
gdown==5.2.0